+++ stapi-fastapi/tests/conftest.py
@@ -176,17 +176,19 @@ def opportunity_search(limit) -> dict[str, Any]:
     end_string = rfc3339_strftime(end, format)
 
     return {
-        "geometry": {
-            "type": "Point",
-            "coordinates": [0, 0],
-        },
-        "datetime": f"{start_string}/{end_string}",
-        "filter": {
-            "op": "and",
-            "args": [
-                {"op": ">", "args": [{"property": "off_nadir"}, 0]},
-                {"op": "<", "args": [{"property": "off_nadir"}, 45]},
-            ],
+        "search_parameters": {
+            "geometry": {
+                "type": "Point",
+                "coordinates": [0, 0],
+            },
+            "datetime": f"{start_string}/{end_string}",
+            "filter": {
+                "op": "and",
+                "args": [
+                    {"op": ">", "args": [{"property": "off_nadir"}, 0]},
+                    {"op": "<", "args": [{"property": "off_nadir"}, 45]},
+                ],
+            },
         },
         "limit": limit,
     }